+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## change log
 
+### FAIR
+- Added `SteamUserStats.getAchievementAndUnlockTime()`.
+- 
 ### [1.10.0-SNAPSHOT]
 - Updated to Steamworks SDK v1.57.
 - Updated Maven modules to compile with Java 8.